Add spec cases for Card class
