+++ spec/game_spec.rb
@@ -1,0 +1,2 @@
+require "rspec"
+require_relative "../models/game"
